Centralize and decouple category state management from bookmarks
Previously, all state and actions for categories were solely coupled to the bookmarks feature. When categories were
added to Applications as well originally all of the bookmark actons/state were reused for proof of concept.  After
a it was kind of working it was realized this made it difficult to reuse the category logic for Applications and
beyond that created a confusion.

This refactor accomplishes the following:
- Creates a new, dedicated 'categories' Redux slice to act as the single source of truth for all category data.
- Moves all category-related action creators and reducer logic from the 'bookmarks' slice to the new 'categories' slice.
- Updates all components (Home, Apps, Bookmarks, and their sub-components) to use the new centralized state and actions.